refactor(net): centralize env proxy in astra-core, fix GitHub client and provider probe policy
Extracts the shared apply_env_proxy helper into astra-core::net so that
the 3e3d6fa8 policy (only external traffic reads env proxy) is enforced
structurally rather than by convention.

Changes:
- astra-core/src/net.rs (new): authoritative pub fn apply_env_proxy with
  policy doc-comment, HTTPS_PROXY/ALL_PROXY/NO_PROXY support, socks5 fix.
  Includes 6 temp_env unit tests (migrated from llm_client.rs).
- astra-core/Cargo.toml: add reqwest + tracing workspace deps; temp-env dev-dep.
- runtime/turn/llm_client.rs: remove local 42-line impl + 84-line test module;
  re-export from astra_core::net instead.
- services/src/models.rs: validate_connectivity: .no_proxy() to apply_env_proxy.
  Provider probes hit external endpoints, not internal services.
- astra-cli/src/edge_tools.rs: GitHub API client: implicit reqwest default to
  apply_env_proxy. Adds TDD regression test (4 proxy env combinations).

Policy invariant enforced:
- External (LLM x2, provider probe, GitHub API): 4 callers use apply_env_proxy
- Internal (memoria, durable, edge, runtime): 24 callers use .no_proxy()
- Test/mock: 3 callers use .no_proxy() or mock

Zero reqwest::Client::builder() calls with implicit proxy behaviour remaining.
