+++ rust/crates/core/src/lib.rs
@@ -14,6 +14,7 @@ pub mod config;
 pub mod drift;
 pub mod error_kind;
 pub mod log;
+pub mod net;
 pub mod runtime_limits;
 
 /// Re-export for [`crate::agent_*!`] macros (call sites do not need a direct `tracing` dependency).
